+++ crates/mysten-common/Cargo.toml
@@ -22,9 +22,6 @@ reqwest.workspace = true
 fastcrypto.workspace = true
 mysten-metrics.workspace = true
 sui-macros.workspace = true
-sui-tls.workspace = true
-sui-types.workspace = true
 tempfile.workspace = true
 tracing.workspace = true
-prometheus.workspace = true
 anyhow.workspace = true
